feat(auth): Extracted common login in App.tsx to useInitializeAuth hook. Removed all usages of useAuthorization. Added some TODOs. Added credentials in all the requests. Removed duplicated hook useSystemInfo
