+++ packages/agents/src/routes/root/BasePage/BasePage.tsx
@@ -19,14 +19,13 @@ import type { Theme } from '@mui/material/styles'
 import type { SystemStyleObject } from '@mui/system/styleFunctionSx/styleFunctionSx'
 import { AppHeader } from '@netcracker/qubership-apihub-ui-shared/components/AppHeader'
 import { MaintenanceNotification } from '@netcracker/qubership-apihub-ui-shared/components/MaintenanceNotification'
-import { SystemInfoPopup } from '@netcracker/qubership-apihub-ui-shared/features/system-info'
+import { SystemInfoPopup, useSystemInfo } from '@netcracker/qubership-apihub-ui-shared/features/system-info'
 import { LogoIcon } from '@netcracker/qubership-apihub-ui-shared/icons/LogoIcon'
 import { cutViewPortStyleCalculator } from '@netcracker/qubership-apihub-ui-shared/utils/themes'
 import type { FC } from 'react'
 import { memo, useCallback } from 'react'
 import { Outlet } from 'react-router-dom'
 import * as packageJson from '../../../../package.json'
-import { useSystemInfo } from '../useSystemInfo'
 import { ErrorNotificationHandler, SuccessNotificationHandler } from './NotificationHandler'
 import { UserPanel } from './UserPanel'
