+++ packages/editor/src/App.tsx
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
+import { ErrorHandler } from '@apihub/components/ErrorHandler'
+import { CssBaseline, ThemeProvider } from '@mui/material'
+import { AppPlaceholder } from '@netcracker/qubership-apihub-ui-shared/components/AppPlaceholder'
+import { useInitializeAuth } from '@netcracker/qubership-apihub-ui-shared/hooks/authorization'
+import { theme } from '@netcracker/qubership-apihub-ui-shared/themes/theme'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import type { FC } from 'react'
 import { memo, StrictMode } from 'react'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { CssBaseline, ThemeProvider } from '@mui/material'
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-import { theme } from '@netcracker/qubership-apihub-ui-shared/themes/theme'
-import { ErrorHandler } from '@apihub/components/ErrorHandler'
 import { Router } from './routes/Router'
 
 const client = new QueryClient({
@@ -35,16 +37,20 @@ const client = new QueryClient({
 })
 
 export const App: FC = memo(() => {
+  const [auth, isLoginPage] = useInitializeAuth()
+
   return (
     <StrictMode>
       <QueryClientProvider client={client}>
         <ThemeProvider theme={theme}>
-          <CssBaseline/>
+          <CssBaseline />
           <ErrorHandler>
-            <Router/>
+            {auth || isLoginPage
+              ? <Router />
+              : <AppPlaceholder />}
           </ErrorHandler>
         </ThemeProvider>
-        <ReactQueryDevtools initialIsOpen={false}/>
+        <ReactQueryDevtools initialIsOpen={false} />
       </QueryClientProvider>
     </StrictMode>
   )
