+++ packages/shared/src/components/AppPlaceholder.tsx
@@ -1,0 +1,9 @@
+import { memo, type FC } from 'react'
+
+export const AppPlaceholder: FC = memo(() => {
+  return (
+    <div style={{ padding: 20, fontSize: 14 }}>
+      Please, wait...
+    </div>
+  )
+})
