+++ packages/shared/src/hooks/authorization/index.ts
@@ -16,4 +16,5 @@
 
 // eslint-disable-next-line filenames/no-index
 export type { AuthorizationOptions, Credentials, LoginUser } from './useUserInfo'
-export { DEFAULT_AUTHORIZATION_DEBOUNCE, useAuthorization } from './useUserInfo'
+export { DEFAULT_AUTHORIZATION_DEBOUNCE, useUserInfo } from './useUserInfo'
+export { useInitializeAuth } from './useInitializeAuth'
